requirements.txt
flask
requests
boto3
snowflake-connector-python
pillow
absl
geopy
swisseph